third_party/blink/web_tests/external/wpt/preload/prefetch-headers.https-expected.txt

This is a testharness.js-based test.
[FAIL] Prefetch should respect CORS mode
  assert_equals: expected (undefined) undefined but got (string) "https://web-platform.test:8444"
Harness: the test ran to completion.
